Minor rewrite to avoid unnecessary updates of frontend, and fixed issue where avg. would not update on match end
